requirements.txt
# --- START OF FILE requirements.txt ---

playwright
beautifulsoup4
discord.py
gunicorn
python-dotenv
Flask